fix: reuse squad wait path and inline worker reports
