+++ packages/cli/src/cli-runtime-wait.ts
@@ -273,6 +273,22 @@ export async function waitForTaskDispatches(command: ThinCommand, taskId: string
   };
 }
 
+export async function waitForSquadRun(command: ThinCommand, squadRunId: string): Promise<JsonObject> {
+  const readCommand = {
+    ...command,
+    method: "repo.task.read",
+    action: { kind: "squad-status", squadRunId },
+  };
+  for (;;) {
+    const status = await runCommandThroughDaemon(readCommand, () => undefined, { autostart: false });
+    if (status.ok !== true) return status;
+    const phase = (status.run as Record<string, unknown> | undefined)?.phase;
+    if (phase === "converged" || phase === "failed" || phase === "cancelled")
+      return { ...status, command: "squad-run", outcome: phase, exitCode: phase === "converged" ? 0 : 1 };
+    await new Promise((resolve) => setTimeout(resolve, 500));
+  }
+}
+
 async function readDaemonSubscription(
   read: () => Promise<JsonObject>,
   reset: () => void = () => undefined,
