+++ packages/daemon/src/squad-leader-decision.ts
@@ -1,4 +1,6 @@
 import type { TaskDispatchRow } from "./protocol/daemon-protocol.contract.ts";
+import { readFileSync } from "node:fs";
+import { resolve } from "node:path";
 
 export type LeaderDecision =
   | { readonly kind: "converged"; readonly report: string | null }
@@ -85,6 +87,7 @@ type LeaderPromptState = {
   readonly roster: string;
   readonly mission: string;
   readonly workerAttempts: readonly WorkerAttempt[];
+  readonly cwd?: string;
 };
 
 export function initialLeaderPrompt(state: LeaderPromptState): string {
@@ -152,9 +155,21 @@ function statusRowsForPrompt(state: LeaderPromptState, rows: readonly TaskDispat
       `exitCode=${String(row?.exitCode ?? "none")}`,
       `resultRef=${row?.resultRef ?? "none"}`,
       `reportPath=${row?.reportPath ?? "none"}`,
+      reportExcerpt(state, row?.reportPath),
       `rejection=${attempt.rejection ?? "none"}`,
     ].join(" ");
   });
+}
+
+function reportExcerpt(state: LeaderPromptState, reportPath: string | null | undefined): string {
+  if (!reportPath) return "reportBody=none";
+  try {
+    const body = readFileSync(resolve(state.cwd ?? process.cwd(), reportPath), "utf8");
+    const excerpt = body.length > 8192 ? `${body.slice(0, 8192)}\n[truncated]` : body;
+    return `reportBody=${excerpt}`;
+  } catch {
+    return "reportBody=unavailable";
+  }
 }
 
 export function parseLeaderDecision(text: string, workers: readonly string[]): LeaderDecision {
